Initial import of existing WIP codebase.
The goal of stan is to make it easier to write project specific
static analysis tests.

Current basic feature set:
 - Parse and type-check packages specified by wildcardable import
   path. Stan re-implements some logic from "go list" so it can walk
   the packages and parse/type-check them at the same time, saving
   time (saved over 2s on internal mono repo).
 - Collate AST info and types info into one *stan.Package object
 - Provide utility method for walking AST w/ list of ancestor
   nodes (often useful).
 - Utility methods for "resolving" exported types and objects by name
   and then searching for them "directly" instead of walking AST. This
   API makes certain types of analysis trivial.
 - Various tricks to try to ensure max code coverage. For example stan
   dedupes symbols in foo_darwin.go and foo_linux.go so both can be
   loaded, type checked (best effort) and ultimately subjected to test
   code.

To be more ergonomic for tests and make it harder to accidentally test
nothing, stan errs on the side of panicking when it can't do/find what
you asked it to.

I have a lot of other ideas I'm going to fool around with. Hopefully
some of them pan out.