Add simple mechanism for writing negative tests.
stan.EvalTest takes a static test function and runs it against some
source code in string form, returning the errors the test
produced. This way it is easy to test that your static tests fail when
they are supposed to.

"static test function" is a function with signature
func(*stan.Package) []error (i.e. a function that operates on
a *stan.Package and returns a list of errors it found performing its
analysis).