Rename Resolve{Type/Object} to Lookup{Type/Object}
"Lookup" is more consistent with naming in standard library. I
changed it to "Resolve" before because the methods "resolve" a type or
object based the invocant package (can differ between packages
depending due to two "worlds" of *types.Package: one from the importer
and one from us invoking the type checker directly). This detail is
not important to the library user, though.

Also shuffle the type checking code into a separate file.